refactor: Move actorinventory browsinginto its own component
- Add ActorInventoryBrowser.svelte
- Make the localization service available to all components
